+++ test/integration/application-ui5-tooling/karma-ui5-config-absolute.conf.js
@@ -1,0 +1,16 @@
+module.exports = function(config) {
+	"use strict";
+
+	require("../karma-base.conf")(config);
+	config.set({
+
+		frameworks: ["ui5"],
+
+		ui5: {
+			configPath: require("path").resolve(__dirname, "ui5-foo.yaml")
+		},
+
+		reporters: ["progress"]
+
+	});
+};